requirements.txt
Django==1.4.2
djangorestframework==2.2.0
django-filter==0.6
django-nose==1.1
coverage==3.6